requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
snowflake-connector-python==3.0.0
plotly==5.17.0
python-dotenv==1.0.0
bcrypt==4.0.1
requests==2.31.0